Splits datastore decorator
