+++ lib/Exceptions/RecordNotFoundException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace PHPNomad\Datastore\Exceptions;
+
+class RecordNotFoundException extends DatastoreErrorException
+{
+}
