+++ lib/Interfaces/Datastore.php
@@ -4,47 +4,13 @@ namespace PHPNomad\Datastore\Interfaces;
 
 use PHPNomad\Datastore\Exceptions\DatastoreErrorException;
 use PHPNomad\Datastore\Exceptions\DuplicateEntryException;
-use React\Dns\RecordNotFoundException;
+use PHPNomad\Datastore\Exceptions\RecordNotFoundException;
 
 /**
  * @template T of <DataModel>
  */
 interface Datastore
 {
-    /**
-     * Query with conditions, using a combination of AND/OR.
-     * Classes implementing this should assume type and groupType are AND if they are not set.
-     *
-     * @param array{type?: string, groupType?: string, clauses: array{column: string, operator: string, value: mixed}[]}[] $conditions
-     * @param positive-int|null $limit
-     * @param positive-int|null $offset
-     * @return T[]
-     * @throws DatastoreErrorException
-     */
-    public function where(array $conditions, ?int $limit = null, ?int $offset = null, ?string $orderBy = null, string $order = 'ASC'): array;
-
-    /**
-     * Query with conditions, using AND.
-     *
-     * @param array{column: string, operator: string, value: mixed}[] $conditions
-     * @param positive-int|null $limit
-     * @param positive-int|null $offset
-     * @return T[]
-     * @throws DatastoreErrorException
-     */
-    public function andWhere(array $conditions, ?int $limit = null, ?int $offset = null, ?string $orderBy = null, string $order = 'ASC'): array;
-
-    /**
-     * Query with conditions, using OR.
-     *
-     * @param array{column: string, operator: string, value: mixed}[] $conditions
-     * @param positive-int|null $limit
-     * @param positive-int|null $offset
-     * @return T[]
-     * @throws DatastoreErrorException
-     */
-    public function orWhere(array $conditions, ?int $limit = null, ?int $offset = null, ?string $orderBy = null, string $order = 'ASC'): array;
-
     /**
      * Finds the first available record that has the specified value in the specified column.
      *
@@ -65,15 +31,6 @@ interface Datastore
      * @throws DatastoreErrorException
      */
     public function create(array $attributes): DataModel;
-
-    /**
-     * Delete a record from the database.
-     *
-     * @param array{string, string} $conditions - field values keyed by their respective field.
-     * @return void
-     * @throws DatastoreErrorException
-     */
-    public function deleteWhere(array $conditions): void;
 
     /**
      * Update a record in the database.
